backend/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-chatbot-backend"
version = "0.1.0"
description = "RAG Chatbot Backend for AI-Spec-Driven Book"
readme = "README.md"
authors = [{name = "Spec-Kit Plus", email = "dev@example.com"}]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "cohere>=4.0.0",
    "qdrant-client>=1.6.0",
    "asyncpg>=0.28.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
    "isort>=5.0"
]

[project.urls]
Homepage = "https://github.com/your-org/rag-chatbot-backend"
Repository = "https://github.com/your-org/rag-chatbot-backend"

[tool.setuptools.packages.find]
where = ["src"]